move old stuff into new structure
