+++ game/src/core/level/generator/graphBased/graph/GraphGenerator.java
@@ -1,0 +1,3 @@
+package core.level.generator.graphBased.graph;
+
+public class GraphGenerator {}
