+++ game/src/core/level/generator/graphBased/graph/GraphLevelGenerator.java
@@ -1,4 +1,4 @@
-package core.level.levelgraph;
+package core.level.generator.graphBased.graph;
 
 import core.level.elements.ILevel;
 import core.level.generator.IGenerator;
